Rename "wordmap" package to "wordmaps"
