+++ internal/wordmaps/mapping_test.go
@@ -1,4 +1,4 @@
-package wordmap
+package wordmaps
 
 import "fmt"
 import "testing"
