+++ internal/wordmaps/wordmap_test.go
@@ -1,4 +1,4 @@
-package wordmap
+package wordmaps
 
 import "testing"
